# openapi.yml
openapi: 3.1.0
info:
  title: Apple Reminders Web API
  description: Apple Reminders 웹 버전 백엔드 API
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: ReminderList
    description: 리마인더 목록 관리
  - name: Reminder
    description: 리마인더 항목 관리

paths:
  /api/reminder-lists:
    get:
      tags: [ReminderList]
      summary: 전체 목록 조회
      description: displayOrder 오름차순으로 모든 리마인더 목록을 반환한다.
      operationId: getAllReminderLists
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReminderListResponse'

    post:
      tags: [ReminderList]
      summary: 목록 생성
      operationId: createReminderList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReminderListRequest'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderListResponse'
        '400':
          description: 유효성 검사 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/reminder-lists/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: 리마인더 목록 ID

    get:
      tags: [ReminderList]
      summary: 단건 조회
      operationId: getReminderListById
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderListResponse'
        '404':
          description: 목록 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [ReminderList]
      summary: 목록 수정
      operationId: updateReminderList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReminderListRequest'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderListResponse'
        '400':
          description: 유효성 검사 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 목록 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [ReminderList]
      summary: 목록 삭제
      operationId: deleteReminderList
      responses:
        '204':
          description: 삭제 성공
        '404':
          description: 목록 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/reminder-lists/{listId}/reminders:
    parameters:
      - name: listId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: 리마인더 목록 ID

    get:
      tags: [Reminder]
      summary: 목록별 리마인더 조회
      description: 특정 목록에 속한 리마인더를 displayOrder 오름차순으로 반환한다.
      operationId: getRemindersByListId
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReminderResponse'
        '404':
          description: 목록 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Reminder]
      summary: 리마인더 생성
      operationId: createReminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderRequest'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '400':
          description: 유효성 검사 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 목록 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    CreateReminderListRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          description: 목록 이름
          example: 장보기
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          default: '#007AFF'
          description: 목록 색상 (hex)
          example: '#FF3B30'

    UpdateReminderListRequest:
      type: object
      required: [name, color, displayOrder]
      properties:
        name:
          type: string
          minLength: 1
          description: 목록 이름
          example: 장보기
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          description: 목록 색상 (hex)
          example: '#FF9500'
        displayOrder:
          type: integer
          description: 사이드바 정렬 순서
          example: 2

    ReminderListResponse:
      type: object
      required: [id, name, color, displayOrder, reminderCount, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
          description: 목록 ID
          example: 1
        name:
          type: string
          description: 목록 이름
          example: 장보기
        color:
          type: string
          description: 목록 색상 (hex)
          example: '#007AFF'
        displayOrder:
          type: integer
          description: 사이드바 정렬 순서
          example: 0
        reminderCount:
          type: integer
          description: 미완료 리마인더 수
          example: 3
        createdAt:
          type: string
          format: date-time
          example: '2026-04-09T10:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2026-04-09T10:30:00'

    ReminderRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          description: 리마인더 제목
          example: 우유 사기

    ReminderResponse:
      type: object
      required: [id, listId, title, isCompleted, displayOrder, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
          description: 리마인더 ID
          example: 1
        listId:
          type: integer
          format: int64
          description: 소속 목록 ID
          example: 1
        title:
          type: string
          description: 리마인더 제목
          example: 우유 사기
        isCompleted:
          type: boolean
          description: 완료 여부
          example: false
        displayOrder:
          type: integer
          description: 표시 순서
          example: 0
        createdAt:
          type: string
          format: date-time
          example: '2026-04-09T10:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2026-04-09T10:30:00'

    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: 에러 메시지
          example: ReminderList not found 99